More accurately calculate exp for exponents above 709.
`f64::exp` for parameters  above ~`709.7827` map to f64::INFINITY.
With this commit, `Q::exp` uses a taylor series with 100 coefficients for parameters >= 709 to avoid this issue.

This taylor series is just generated once and then stored for later use. For this, the Sync trait had to be implemented for PolyOverQ. Since Send also applies, it is now also implemented.